requirements.txt
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
DateTime==4.3
Flask==2.0.2
Flask-MySQLdb==0.2.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysqlclient==2.1.0
Pillow==9.0.0
pycparser==2.21
pytz==2021.3
six==1.16.0
stepic==0.5.0
Werkzeug==2.0.2
zope.interface==5.4.0
